feat(pages): organize page comps into public and private
