+++ src/App.tsx
@@ -7,10 +7,11 @@ import { Store } from "redux";
 
 import Header from "./components/Header";
 import client from "./graphql/client";
+import { PrivateRoutesRegistry } from "./pages/private";
 import {
   LoginPage,
-  PrivateRoutesRegistry,
-} from "./pages";
+  WipChallengesPage,
+} from "./pages/public";
 import { history } from "./redux/ConfigureStore";
 
 export default class App extends React.Component<{ store: Store }, {}> {
@@ -19,6 +20,8 @@ export default class App extends React.Component<{ store: Store }, {}> {
       <ConnectedRouter history={history}>
         <ApolloProvider client={client(this.props.store)}>
           <Header />
+          <Route exact path="/" component={WipChallengesPage} />
+          <Route path="/wip_challenges" component={WipChallengesPage} />
           <Route path="/login" component={LoginPage} />
           <PrivateRoutesRegistry/>
         </ApolloProvider>
