+++ src/pages/private/PrivateRoutesRegistry.tsx
@@ -3,15 +3,12 @@ import {
   ChallengesPage,
   EditChallengePage,
   NewChallengePage,
-  WipChallengesPage,
 } from ".";
-import PrivateRoute from "../components/PrivateRoute";
+import PrivateRoute from "../../components/PrivateRoute";
 
 export default function PrivateRoutesRegistry() {
   return (
     <React.Fragment>
-      <PrivateRoute exact path="/" component={WipChallengesPage} />
-      <PrivateRoute path="/wip_challenges" component={WipChallengesPage} />
       <PrivateRoute
         exact
         path="/challenges"
