+++ src/pages/private/index.ts
@@ -1,15 +1,11 @@
 import ChallengesPage from "./ChallengesPage";
 import EditChallengePage from "./EditChallengePage";
-import LoginPage from "./LoginPage";
 import NewChallengePage from "./NewChallengePage";
 import PrivateRoutesRegistry from "./PrivateRoutesRegistry";
-import WipChallengesPage from "./WipChallengesPage";
 
 export {
   ChallengesPage,
   EditChallengePage,
-  LoginPage,
   NewChallengePage,
   PrivateRoutesRegistry,
-  WipChallengesPage,
 };
