+++ src/pages/public/index.ts
@@ -1,0 +1,7 @@
+import LoginPage from "./LoginPage";
+import WipChallengesPage from "./WipChallengesPage";
+
+export {
+  LoginPage,
+  WipChallengesPage,
+};
