+++ src/styles/pages/ChallengesPage.css
